requirements.txt
django==4.0
psycopg2-binary==2.9.7
djangorestframework==3.14.0
markdown==3.4.4
django-filter==23.2
drf_yasg==1.21.7